refactor(swipe): move to script setup
