+++ src/packages/__VUE/addresslist/components/GeneralShell.taro.vue
@@ -47,7 +47,7 @@ import { createComponent } from '@/packages/utils/create';
 const { create } = createComponent('address-list-general');
 import ItemContents from './ItemContents.taro.vue';
 import NutButton from '../../button/index.taro';
-import NutSwipe from '../../swipe/index.taro.vue';
+import NutSwipe from '../../swipe/index.taro';
 
 export default create({
   props: {
