+++ src/packages/__VUE/swipe/types.ts
@@ -1,1 +1,2 @@
 export const SWIPE_KEY = Symbol('nut-swipe');
+export type SwipePosition = 'left' | 'right' | '';
